+++ cantera/trunk/src/thermo/SpeciesThermoInterpType.cpp
@@ -57,7 +57,7 @@ doublereal SpeciesThermoInterpType::reportHf298(doublereal* const h298) const
 }
 //====================================================================================================
 
-void SpeciesThermoInterpType::modifyOneHf298(const int k, const doublereal Hf298New)
+void SpeciesThermoInterpType::modifyOneHf298(const size_t k, const doublereal Hf298New)
 {
     throw CanteraError("SpeciesThermoInterpType::modifyOneHf298",
                        "Not implemented");
